Moved The full value sentence from A to Keys
Moved "The full value..." sentence from the A shared inheritance section to the beginning of the Keys to success section. Edited the leads of the Taxpayers sentence, to reflect it now follows the full value sentence. Specified that taxpayers, "must have an easy to use tool enabling them to track where their tax dollars go..."