Changed 'our' to 'a'; replaced em dashes
In Choosing section, at the end: changed "our shared inheritance" to "a shared inheritance". Replaced first em dash with a comma in the final sentence; replaced the last em dash with a colon. Removed the word 'such'; changed the word 'our' to 'those' in referencing 'decisions'. Inserted the word 'civilized' in between 'to' & 'living'