Replaced sentence in How with a new paragraph
In the "How..." section: broke off the last sentence into a paragraph and changed "Classical Liberals seek to remove forms of privilege, especially that privilege enabled by the state." to "For the last half millennium Classical Liberals have been working to remove state enabled privilege. It's the hope of this author that expanding the recognized rights of [wo]men to include an explicit portion of nature universally would resolve not only conflicts between people but also enable the same to always have a fallback position in any negotiation. For more on this set see https://bigthink.com/thinking/classical-liberalism-explained/."