Updated With section
In "With regard to government" section: Broke last sentence off to be it's own standalone. Changed the sentence from "The best government is the one where decisions are made at the most local competent level possible." to "In accordance to the principle of Subsidiarity: the best government is the one where decisions are made at the most local competent level possible." Migrated the hyperlinked word in the sentence from 'decisions' to 'Subsidiarity'.