Update README.md
Under the Priorities section: removed the text highlight from the pigovian hyperlink target because the updated version of the page no longer has the referenced text strings by which to highlight. Plus, given the page's new design–which includes a short definition in a graphic, the former highlight is no longer necessary.